master/requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
branca==0.4.1
Brlapi==0.7.0
buttercms-python==1.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.2.1
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dicttoxml==1.7.4
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.3
django-appconf==1.0.4
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-easy-maps==1.1.2
django-extensions==3.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-leaflet==0.27.1
django-phonenumber-field==5.0.0
django-ranged-response==0.2.0
django-sendsms==0.4
django-simple-captcha==0.5.13
django-taggit==1.3.0
Django-Verify-Email==0.0.4
django-wordpress-api==0.2.0
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
folium==0.11.0
freeze==3.0
future==0.18.2
GDAL==3.0.4
geocoder==1.38.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.0.0
graphviz==0.14.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
ip2geotools==0.1.5
IP2Location==8.5.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kavenegar==1.1.2
keyring==21.5.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
leaflet==0.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
multidict==5.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.13
pickleshare==0.7.5
Pillow==8.0.1
pip-review==1.1.0
pkginfo==1.6.1
postgis==1.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.6.1
protonvpn-cli==2.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pydotplus==2.0.2
Pygments==2.7.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyquery==1.4.3
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
Python-dev==2.0.0.dev0
pythondialog==3.5.1
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
ratelim==0.1.6
readme-renderer==28.0
redis==3.5.3
reportlab==3.5.34
requests==2.21.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.4.0
SecretStorage==3.3.0
selenium==3.141.0
Send2Trash==1.5.0
sendsms==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.4.1
systemd-python==234
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
twilio==6.50.1
twine==3.2.0
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
usb-creator==0.3.7
validate-email==1.3
virtualenv==20.1.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
